requirements.txt
requests>=2.32.0,<3
python-dotenv>=1.0.1
sqlalchemy>=2.0.30,<3
